// ldk/node/dist/loopClients/whisperSensor.d.ts
export declare enum WhisperListStyle {
    NONE = 0,
    SUCCESS = 1,
    WARN = 2,
    ERROR = 3
}
export declare enum WhisperListAlign {
    LEFT = 0,
    CENTER = 1,
    RIGHT = 2
}